+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Add `SubjectProfile` resource
+- Add `AccountSettings.auditLogSubjectProfile` property
 - Add `NotificationChannels.afterSent` property
 - Add `NotificationChannels.envelopeNotification` property
 - Add `AccountBilling.bankIdSign` property